+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- fix a lot of pytest-asyncio problems in the unit tests  
 - Cleanup startup code for Quart and the Proxy
 - Redirect the hypercorn traces to a separate log-file
 - Configure the dashboard trace handler by the logging.ini file